split convert family instance into several files
