+++ Objects/Objects/BuiltElements/Revit/Interfaces/IHasRevitSymbolType.cs
@@ -1,0 +1,7 @@
+namespace Objects.BuiltElements.Revit.Interfaces
+{
+  public interface IHasRevitSymbolType
+  {
+    string type { get; }
+  }
+}
