+++ Objects/Objects/Objects.csproj
@@ -14,10 +14,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Tests"/>
+    <InternalsVisibleTo Include="Tests" />
   </ItemGroup>
   
   <ItemGroup>
-    <ProjectReference Include="..\..\Core\Core\Core.csproj" Private="false"/>
+    <ProjectReference Include="..\..\Core\Core\Core.csproj" Private="false" />
   </ItemGroup>
 </Project>
